packages/privacy/pyproject.toml
[project]
name = "openadapt-privacy"
version = "0.1.0"
description = "Privacy scrubbing for GUI automation data - PII/PHI detection and redaction"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "MLDSAI Inc.", email = "richard@mldsai.com"}
]
keywords = ["privacy", "pii", "phi", "scrubbing", "redaction", "gui", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pillow>=10.0.0",
    "pip>=25.3",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
presidio = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "presidio-image-redactor>=0.0.50",
    "spacy>=3.7.0",
    "spacy-transformers>=1.3.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/OpenAdaptAI/openadapt-privacy"
Repository = "https://github.com/OpenAdaptAI/openadapt-privacy"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openadapt_privacy"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
